Force unfreeze when the minNumberOfPlayers-th player leaves during a blood moon.
During a blood moon, it's not desirable that when the
minNumberOfPlayers-th player disconnects, the blood moon is cancelled
and reset to the previous day.

In this case, when the last player disconnects, immediately force
unfreeze the time. This can only be cancelled using stoptime or by
everyone leaving the server (and later rejoining).